Added a fullpage for success message for recover account
